Support VS2019 builds
Do not merge!

This may require additional work to sort through necessary prior
commits, like the Scons3 work.

The included new python files in

    omaha/site_scons/site_tools/*_vc16_*.py

are almost entirely derivative of prior versions of those files,
so the copyrights are copied unchanged.

Additional breakpad fixes are necessary, but might be resolved
with more warnings disabled.