Remove gates that are causing a deadlock.
The AppBundle destructor was using a gate to synchronize with another thread that was sending a ping. If the AppBundle destructor was being called in the context of a ::CoUninitialize, this was causing a deadlock, because the new thread called ::CoInitializeEx, which waited on a lock with ::CoUninitialize.

The change is to Lock the ATL Module before creating the thread, and Unlocking the ATL Module at the end of the thread proc.

Fixed a similar potential issue in OnDemand.