Fix broken unit test CertInfoTest.CertInfo.
The SHA1 codesigning certificate for Google has changed.

SHA1 code signing is needed for XP support (which Google Update needs)
but it is not needed for the open source upstream. However, code
changes were needed to support a SHA1 certificate with a different
subject name (Google Inc vs. Google LLC).