Add test case to url_utils.
This commit fixes the unit test as well.

[ RUN      ] BuildQueryStringTest.EncodesQueryElements
common\url_utils_unittest.cc(40): error: Expected equality of these values:
  query
    Which is: L"first=%20'%60&second=%26="
  L"first=%7F%91%80&second=%26%3D"
[  FAILED  ] BuildQueryStringTest.EncodesQueryElements (1 ms)
[----------] 1 test from BuildQueryStringTest (3 ms total)

The underlying code which is called to url escape strings is not fully
RFC-standard and it ignores escaping some characters. It is not agreed
at this time whether the production code to be made more compliant or
just leave it as-is. This commit just merges in a unit test and makes
the test pass with the existing production code.