Remove gates that could potentially cause a deadlock.
Removed the gate in CoCreateAsyncStatus::CreateOmahaMachineServerAsync that was used to synchronize with another thread. This is because if CoCreateAsyncStatus::CreateOmahaMachineServerAsync is called in the context of a ::CoUninitialize, this code would have caused a deadlock, because the new thread calls ::CoInitializeEx, which waits on a lock with ::CoUninitialize.

The change is to AddRef the Object before creating the thread, and Releasing the object at the end of the thread proc.