Restrict ACLs on the crash handler named pipe.
This CL changes from passing in a more permissive security descriptor
to NULL. The ACLs in the default security descriptor for a named pipe
grant full control to the LocalSystem account, administrators, and the
creator owner. This is sufficient for now, since Omaha is the only
client for the corresponding Breakpad server.

The drawback is that we will not be able to capture crashes that
happen in Medium Integrity Machine Omaha processes, which is a corner
case, and hence is ok.