Security: Local Elevation of Privilege vulnerability in Google Update Service.
Summary of the security issue: An attacker can control the location and filename
where GoogleUpdate logs information if they can write a GoogleUpdate.ini under C:\.

To mitigate this issue, this change does the following:
* removes the ability to specify a `LogFilePath` in GoogleUpdate.ini. This implies
that the log file will always be GoogleUpdate.log and always under
the %ALLUSERSPROFILE%\Google\Update\Log directory.
* checks whether the file is a reparse point after opening the file. The check
is made after opening the file, so that the attacker does not get a chance to
substitute a reparse point.
* as a defense in depth measure, we check to make sure the parent directory
has not been redirected. i.e., the %ALLUSERSPROFILE%\Google\Update directory.
We do not check %ALLUSERSPROFILE%\Google and above for reparse points, since
an attacker would need to reuse an existing directory structure which has "\Update",
which narrows the attack surface considerably, and in addition, we only write to a
"GoogleUpdate.log" file within, which is unlikely to affect most applications
(such as GoogleUpdate, which has that directory structure under %ProgramFiles (x86)%).

crbug.com/1204811.