Disable calling ::EnterCriticalPolicySection/::LeaveCriticalPolicySection
We are getting reports from the field that GoogleUpdate gets  deadlocked on EnterCriticalPolicySection when the installer is run via GPO.
https://bugs.chromium.org/p/chromium/issues/detail?id=1231513#c13
It appears that when the install is pushed via GPO, GPO takes the critical section, and we hang at the ::EnterCriticalPolicySection(TRUE) call. So, for now, we are going to just dismiss the guard.